requirements.txt
openpyxl
pillow
selenium